Split figures counter declarations and definitions
